Refactor BaseModel to db utilities
