+++ commcare_connect/utils/db.py
@@ -1,4 +1,15 @@
+from django.db import models
 from django.utils.text import slugify
+
+
+class BaseModel(models.Model):
+    created_by = models.CharField(max_length=255)
+    modified_by = models.CharField(max_length=255)
+    date_created = models.DateTimeField(auto_now_add=True)
+    date_modified = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        abstract = True
 
 
 def slugify_uniquely(value, model, slugfield="slug"):
